+++ node/utils/object.ts
@@ -17,11 +17,12 @@ import { adjust, curry, fromPairs, map, mergeAll, pipe, toPairs, zipObj } from '
  *
  * renameKeysWith(concat('a'), { A: 1, B: 2, C: 3 }) //=> { aA: 1, aB: 2, aC: 3 }
  */
+type Tuple = [string, any]
 const renameKeysWith = curry((func: any, object: any) =>
-  pipe(
+  pipe<any, Tuple[], any, any>(
     toPairs,
-    map(adjust(func, 0) as any),
-    fromPairs as any
+    map(adjust(0, func)),
+    fromPairs
   )(object)
 )