+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.113.2] - 2020-01-06
 ### Changed
 - Use native promises and not bluebird anymore.
 - Other code improvements related to types and ramda update.